+++ packages/textfield/src/component.tsx
@@ -1,8 +1,13 @@
 import React, { forwardRef } from 'react';
+import { i18n } from '@lingui/core';
 import { classNames } from '@chbphone55/classnames';
 import { input as ccInput, label as ccLabel, helpText as ccHelpText } from '@warp-ds/css/component-classes';
 import { useId } from '../../utils/src';
 import { TextFieldProps } from './props';
+import { messages as nbMessages} from './locales/nb/messages.mjs';
+import { messages as enMessages} from './locales/en/messages.mjs';
+import { messages as fiMessages} from './locales/fi/messages.mjs';
+import { activateI18n } from '../../i18n';
 
 export const TextField = forwardRef<HTMLInputElement, TextFieldProps>(
   (props, ref) => {
@@ -21,6 +26,8 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>(
       optional,
       ...rest
     } = props;
+
+    activateI18n(enMessages, nbMessages, fiMessages);
 
     const id = useId(providedId);
 
@@ -47,7 +54,13 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>(
               {label}
               {optional && (
                 <span className={ccLabel.optional}>
-                  (valgfritt)
+                  {i18n._(
+                    /*i18n*/ {
+                      id: 'textfield.label.optional',
+                      message: '(optional)',
+                      comment: 'Shown behind label when marked as optional',
+                    },
+                  )}
                 </span>
               )}
             </label>